refactor: address PR review feedback
- ui-react: extract the splat base-path math into a shared useSplatBasePath
  hook; use it in both GSPageLayout and MusterSection (was duplicated).
- ui-react: useProvidePageHeaderActions now depends on [actions], and the
  agent create/review pages memoize their actions element, so the section
  header no longer re-renders on every keystroke.
- app: document that agentPlatformPlugin must be registered before musterPlugin
  (feature order determines the level-1 tab order).
- muster: note that MusterSection's px="5" gutter is hand-matched to bui's
  PluginHeader/Content padding.
