+++ plugins/ui-react/src/hooks/index.ts
@@ -1,4 +1,5 @@
 export * from './useContainerDimensions';
 export * from './useDetailsPane';
 export * from './useFilters';
+export * from './useSplatBasePath';
 export * from './useTableColumns';
